Begin implementing liblocus
Liblocus is a part of libmailutils that will provide functions for
manipulating source file locations, for use in lexers, grammars, etc.
This will expand the functionality of the mu_locus type and logstreams.

* configure.ac: add libmailutils/locus/
* include/mailutils/assoc.h (mu_assoc_install_ref2): New proto.
* libmailutils/base/assoc.c (mu_assoc_install_ref2): New entry point.
* libmailutils/base/copyfile.c (copy_regular_file): Add typecasts.
* libmailutils/Makefile.am: Build liblocus
* libmailutils/locus/Makefile.am: New file.
* libmailutils/locus/debug.c: New file.
* libmailutils/locus/ident.c: New file.
* libmailutils/locus/tracker.c: New file.
* libmailutils/tests/Makefile.am: New file.
* libmailutils/tests/tracker.c: New file.

+++ libmailutils/base/copyfile.c
@@ -163,7 +163,7 @@ copy_regular_file (const char *srcpath, const char *dstpath, int flags,
       rc = mu_stream_ioctl (dst, MU_IOCTL_TRANSPORT, MU_IOCTL_OP_GET, trans);
       if (rc == 0)
 	{	    
-	  if (fchmod ((int) trans[0], mode))
+	  if (fchmod ((int) (intptr_t) trans[0], mode))
 	    {
 	      mu_debug (MU_DEBCAT_STREAM, MU_DEBUG_ERROR,
 			(_("%s: cannot chmod: %s"),
@@ -193,7 +193,7 @@ copy_regular_file (const char *srcpath, const char *dstpath, int flags,
 
 	      if (gid != -1)
 		{
-		  if (fchown ((int) trans[0], uid, gid))
+		  if (fchown ((int) (intptr_t) trans[0], uid, gid))
 		    {
 		      mu_debug (MU_DEBCAT_STREAM, MU_DEBUG_ERROR,
 				(_("%s: cannot chown to %lu.%lu: %s"),
